+++ src/Sentry.Unity/Metrics/IGameMetricMonitor.cs
@@ -1,0 +1,6 @@
+namespace Sentry.Unity.Metrics;
+
+internal interface IGameMetricMonitor
+{
+    void Sample();
+}